+++ Frends.AzureDataLake.DeleteContainer/Frends.AzureDataLake.DeleteContainer/Definitions/Options.cs
@@ -1,4 +1,5 @@
 using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
 
 namespace Frends.AzureDataLake.DeleteContainer.Definitions;
 
@@ -8,9 +9,18 @@ namespace Frends.AzureDataLake.DeleteContainer.Definitions;
 public class Options
 {
     /// <summary>
-    /// Throw an error if container to be deleted doesn't exists.
+    /// Throw an exception if the task encounters an error.
+    /// When false, the task returns a Result with Success = false and Error details instead.
     /// </summary>
-    /// <example>false</example>
-    [DefaultValue(false)]
-    public bool ThrowErrorIfContainerDoesNotExist { get; set; }
+    /// <example>true</example>
+    [DefaultValue(true)]
+    public bool ThrowErrorOnFailure { get; set; } = true;
+
+    /// <summary>
+    /// Optional custom error message to use when ThrowErrorOnFailure is true.
+    /// </summary>
+    /// <example></example>
+    [DisplayFormat(DataFormatString = "Text")]
+    [DefaultValue("")]
+    public string ErrorMessageOnFailure { get; set; } = string.Empty;
 }